add empty inventory id filter

+++ service/static/js/rest_api.js
@@ -241,22 +241,27 @@ $(function () {
     $("#delete-btn").click(function () {
         var inventory_id = $("#inventory_id").val();
 
-        var ajax = $.ajax({
-            type: "DELETE",
-            url: "/inventory/" + inventory_id,
-            contentType: "application/json",
-            data: '',
-        })
-
-        ajax.done(function(res){
+        if(!inventory_id.trim().length) {
             clear_form_data()
-            flash_message("Inventory "+ inventory_id +" has been Deleted!")
-            list_all_inventories()
-        });
+            flash_message("Inventory Id should not be empty!")
+        } else {
+            var ajax = $.ajax({
+                type: "DELETE",
+                url: "/inventory/" + inventory_id,
+                contentType: "application/json",
+                data: '',
+            })
 
-        ajax.fail(function(res){
-            flash_message("Server error!")
-        });
+            ajax.done(function(res){
+                clear_form_data()
+                flash_message("Inventory "+ inventory_id +" has been Deleted!")
+                list_all_inventories()
+            });
+
+            ajax.fail(function(res){
+                flash_message(res.responseJSON.message)
+            });
+        }
     });
 
     // ****************************************
